Added implicit constructor for a student object. Added assummptions object.
The object RulesAndAssumptions contains function that rely heavily on convention.
This is a temporary measure to aggregate all computation that would otherwise rely on
implicit assumptions. When a method is devised of distributing these functions to their
appropriate context yet make it easy for anyone who reads the code to immediately know
what the major assumptions of the program are then this class needs to be factored out asap.

Note: this is something that I have never tried before. I am curious how it will turn out.