Added functionality pertaining to Homework info.
All homeworks are represented by a homework metadata file that contains
information about the test suites to be run, what imports are required
to run the tests and what the points corresponding to each test-suite.
This information is used to generate a boilerplate test file that will
be run by this program to gather the data necessary to determine how the
students fared.

Added parsing of such files.

Added basic implementation of how homework grading should be done.