Handled non-termination of homework assignments.
Racket processes that last more than 20s are now killed and
and a humorous line is appended to the processes stderr output.

This is done by spawning an actor that waits for 20seconds to receive
a message indicating that all is fine. If the message is not received
then it goes ahead and queries the processes running, it looks for the
`full-path-to-racket` process and it runs a `kill -9 $pid` command.

This solution was chosen because the Java api is annoying to work with.
However, that change will have to be done sooner or later. For now, this
will do.