Added a `create-report` command
It will create a report for the currently loaded homework info and the currently
loaded students. It looks on the file system for the feedback.md files. It looks for a
specific line that was added to the feedback file template only after homework02.
So creating reports before that would not work.